feat: implement the external adc

+++ board/rriv_board/src/lib.rs
@@ -95,7 +95,7 @@ pub trait SensorDriverServices {
     // fn query adc by index
     
     fn query_internal_adc(&mut self, port: u8) -> u16;
-    fn query_external_adc(&mut self, port: u8) -> u32;
+    fn query_external_adc(&mut self, port: u8) -> u16;
     fn ic2_read(&mut self, addr: u8, buffer: &mut [u8]) -> Result<(), ()>;
     fn ic2_write(&mut self, addr: u8, message: &[u8]) -> Result<(), ()>;
